+++ python/dartpy/dynamics/BodyNode.cpp
@@ -968,10 +968,14 @@ void BodyNode(py::module& m)
           })
       .def(
           "getChainDofs",
-          +[](const dart::dynamics::BodyNode* self)
-              -> const std::vector<const dart::dynamics::DegreeOfFreedom*> {
-            return self->getChainDofs();
-          })
+          +[](dart::dynamics::BodyNode* self)
+              -> std::vector<dart::dynamics::DegreeOfFreedom*> {
+            std::vector<dart::dynamics::DegreeOfFreedom*> dofs;
+            for (const auto* dof : self->getChainDofs())
+              dofs.push_back(const_cast<dart::dynamics::DegreeOfFreedom*>(dof));
+            return dofs;
+          },
+          ::py::return_value_policy::reference_internal)
       .def(
           "addExtForce",
           +[](dart::dynamics::BodyNode* self, const Eigen::Vector3d& _force) {