+++ python/dartpy/dynamics/Skeleton.cpp
@@ -527,10 +527,11 @@ void Skeleton(py::module& m)
           ::py::arg("name"))
       .def(
           "getDofs",
-          +[](const dart::dynamics::Skeleton* self)
-              -> std::vector<const dart::dynamics::DegreeOfFreedom*> {
+          +[](dart::dynamics::Skeleton* self)
+              -> std::vector<dart::dynamics::DegreeOfFreedom*> {
             return self->getDofs();
-          })
+          },
+          ::py::return_value_policy::reference_internal)
       .def(
           "getIndexOf",
           +[](const dart::dynamics::Skeleton* self,